# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  username: string;
}

export interface AuthSession {
  authenticated: boolean;
  user: AuthUser | null;
}

export interface AuthMessage {
  message: string;
  /** @nullable */
  email?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface SignInInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SignUpInput {
  /** @minLength 1 */
  username: string;
  /** Valid email address */
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface VerifyEmailInput {
  /** Valid email address */
  email: string;
  /** @pattern ^[0-9]{6}$ */
  token: string;
}

export interface ResendVerificationInput {
  /** Valid email address */
  email: string;
}