# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the current authenticated session
      responses:
        "200":
          description: Current session, or an unauthenticated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/sign-in:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in with a username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInInput"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid credentials or input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/sign-up:
    post:
      operationId: signUp
      tags: [auth]
      summary: Register a new Outfy user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpInput"
      responses:
        "201":
          description: Registration accepted and email verification required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "400":
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify an email with a six-digit code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Verified and authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-code:
    post:
      operationId: resendVerificationCode
      tags: [auth]
      summary: Resend the email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Verification email requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/sign-out:
    post:
      operationId: signOut
      tags: [auth]
      summary: Sign out the current user
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
      required:
        - id
        - username
    AuthSession:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required:
        - authenticated
        - user
    AuthMessage:
      type: object
      properties:
        message:
          type: string
        email:
          type: ["string", "null"]
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SignInInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - username
        - password
    SignUpInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        email:
          type: string
          description: Valid email address
        password:
          type: string
          minLength: 8
      required:
        - username
        - email
        - password
    VerifyEmailInput:
      type: object
      properties:
        email:
          type: string
          description: Valid email address
        token:
          type: string
          pattern: "^[0-9]{6}$"
      required:
        - email
        - token
    ResendVerificationInput:
      type: object
      properties:
        email:
          type: string
          description: Valid email address
      required:
        - email