# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  /** Valid email address */
  email: string;
  /** @pattern ^[0-9]{6}$ */
  token: string;
}